requirements.txt
# Dev/Deployment
pyyaml